Checking the path of the loading matrix

Makes sure that the loading matrix is located at the filepath stored in path2load_mat

Usage

checkLoadingMat(path2load_mat)

Arguments

path2load_mat

[character] /path/to/directory where loading matrices are kept. This can be downloaded from the github repo. Note that the name of the file before the .eigenvec.allele or .acount must be included in file path.

Examples

path2load_mat <- '/path/to/loading_mat/merged_chrs.postQC.train.pca'
if (FALSE) { # \dontrun{
# the following code is not run on package build, as the path2load_mat on the
# user system is not known.
checkLoadingMat(path2load_mat = path2load_mat)
} # }