Identification of individuals of divergent ancestry

Evaluates and depicts results of plink --pca on merged genotypes from individuals to be QCed and individuals of reference population of known genotypes. Currently, check ancestry only supports automatic selection of individuals of European descent. It uses information from principal components 1 and 2 returned by plnik --pca to find the center of the European reference samples (mean(PC1_europeanRef), mean(PC2_europeanRef). It then computes the maximum euclidean distance (maxDist) of the European reference samples from this centre. All study samples whose euclidean distance from the centre falls outside the circle described by the radius r=europeanTh* maxDist are considered non-European and their IDs are returned as failing the ancestry check. check_ancestry creates a scatter plot of PC1 versus PC2 color-coded for samples of the reference populations and the study population.

check_ancestry(qcdir, alg, prefixMergedDataset, europeanTh = 1.5,
  refSamples = NULL, refColors = NULL, refSamplesFile = NULL,
  refColorsFile = NULL, refSamplesIID = "IID", refSamplesPop = "Pop",
  refColorsColor = "Color", refColorsPop = "Pop",
  studyColor = "#2c7bb6", interactive = FALSE)

Arguments

qcdir
[character] path/to/directory/with/QC/results containing prefixMergedDataset.eigenvec results as returned by plink --pca. User needs writing permission to qcdir.

alg
[character] prefix of QC-ed plink files, i.e. alg.bed, alg.bim, alg.fam.

prefixMergedDataset
[character] Prefix of merged dataset (study and reference samples) used in plink --pca, resulting in prefixMergedDataset.eigenvec

europeanTh
[double] Scaling factor of radius to be drawn around center of reference European samples, with study samples inside this radius considered to be of European descent and samples outside this radius of non-European descent. The radius is computed as the maximum Euclidean distance of reference Europeans to the centre of reference European samples.

refSamples
[data.frame] Dataframe with sample identifiers [refSamplesIID] corresponding to IIDs in prefixMergedDataset.eigenvec and population identifier [refSamplesPop] corresponding to population IDs [refColorsPop] in refColorsfile/refColors. Either refSamples or refSamplesFile have to be specified.

refColors
[data.frame, optional] Dataframe with population IDs in column [refColorsPop] and corresponding color-code for PCA plot in column [refColorsColor]. If not provided and is.null(refColorsFile) default colors are used.

refSamplesFile
[character] /Path/to/File/with/reference samples. Needs columns with sample identifiers [refSamplesIID] corresponding to IIDs in prefixMergedDataset.eigenvec and population identifier [refSamplesPop] corresponding to population IDs [refColorsPop] in refColorsfile/refColors.

refColorsFile
[character, optional] /Path/to/File/with/Population/Colors cotaining population IDs in column [refColorsPop] and corresponding color-code for PCA plot in column [refColorsColor].If not provided and is.null(refColors) default colors for are used.

refSamplesIID
[character] Column name of reference sample IDs in refSamples/refSamplesFile.

refSamplesPop
[character] Column name of reference sample population IDs in refSamples/refSamplesFile.

refColorsColor
[character] Column name of population colors in refColors/refColorsFile

refColorsPop
[character] Column name of reference sample population IDs in refColors/refColorsFile.

studyColor
[character] Color to be used for study population if plot is TRUE.

interactive
[logical] Should plots be shown interactively? When choosing this option, make sure you have X-forwarding/graphical interface available for interactive plotting. Alternatively, set interactive=FALSE and save the returned plot object (p_ancestry) via ggplot2::ggsave(p=p_ancestry, other_arguments) or pdf(outfile) print(p_ancestry) dev.off().

Value

named [list] with i) fail_ancestry, containing a [data.frame] with FID and IID of non-European individuals and ii) p_ancestry, a ggplot2-object 'containing' a scatter plot of PC1 versus PC2 color-coded for samples of the reference populations and the study population, which can be shown by print(p_ancestry).