Identification of SNPs showing a significant deviation from Hardy-Weinberg- equilibrium (HWE)

Uses plink --remove alg.fail.IDs --hwe to calculate the observed and expected heterozygote frequencies per SNP in the individuals that passed the perIndividualQC and compute the deviation of the frequencies from Hardy-Weinberg equilibrium (HWE) by HWE exact test. The p-values of the HWE exact test are displayed as histograms (stratified by all and low p-values), where the hweTh is used to depict the QC cut-off for SNPs.

check_hwe(qcdir, alg, hweTh = 1e-05, interactive = FALSE,
  path2plink = NULL, verbose = FALSE, showPlinkOutput = TRUE)

Arguments

qcdir
[character] /path/to/directory/with/QC/results containing alg.bim, alg.bed, alg.fam files and alg.fail.IDs containing IIDs of individuals that failed QC. User needs writing permission to qcdir.

alg
[character] prefix of QC-ed plink files, i.e. alg.bed, alg.bim, alg.fam.

hweTh
[double] Significance threshold for deviation from HWE.

interactive
[logical] Should plots be shown interactively? When choosing this option, make sure you have X-forwarding/graphical interface available for interactive plotting. Alternatively, set interactive=FALSE and save the returned plot object (p_hwe) via ggplot2::ggsave(p=p_hwe, other_arguments) or pdf(outfile) print(p_hwe) dev.off().

path2plink
[character] Absolute path to where external plink software https://www.cog-genomics.org/plink/1.9/ can be found. If not provided, assumed that PATH set-up works and plink will be found by system("plink").

verbose
[logical] If TRUE, progress info is printed to standard out and specifically, if TRUE, plink log will be displayed.

showPlinkOutput
[logical] If TRUE, plink log and error messages are printed to standard out.

Value

named list with i) fail_hwe containing a [data.frame] with CHR (Chromosome code), SNP (Variant identifier), TEST (Type of test: one of 'ALL', 'AFF', 'UNAFF', 'ALL(QT)', 'ALL(NP)'), A1 (Allele 1; usually minor), A2 (Allele 2; usually major), GENO ('/'-separated genotype counts: A1 hom, het, A2 hom), O(HET) (Observed heterozygote frequency E(HET) (Expected heterozygote frequency), P (Hardy-Weinberg equilibrium exact test p-value) for all SNPs that failed the hweTh and ii) p_hwe, a ggplot2-object 'containing' the HWE p-value distribution histogram which can be shown by (print(p_hwe)).