Create plink dataset with samples and markers passing quality control

Remove samples that fail per-sample QC and remove markers that fail per-marker QC.

cleanData(qcdir, alg, lmissTh = 0.01, hweTh = 1e-05, macTh = 20,
  mafTh = 0.01, filterAncestry = TRUE, filterRelated = TRUE,
  filterSex = TRUE, filterHeterozygosity = TRUE,
  filterSampleMissingness = TRUE, filterSNPMissingness = TRUE,
  filterHWE = TRUE, filterMAF = TRUE, path2plink = NULL,
  verbose = FALSE, showPlinkOutput = TRUE)

Arguments

qcdir
[character] /path/to/directory/with/QC/results containing alg.bim, alg.bed, alg.fam files and alg.fail.IDs containing IIDs of individuals that failed QC. User needs writing permission to qcdir.

alg
[character] prefix of QC-ed plink files, i.e. alg.bed, alg.bim, alg.fam.

lmissTh
[double] Threshold for acceptable variant missing rate across samples.

hweTh
[double] Significance threshold for deviation from HWE.

macTh
[double] Threshold for minor allele cut cut-off, if both mafTh and macTh are specified, macTh is used (macTh = mafTh\*2\*NrSamples)

mafTh
[double] Threshold for minor allele frequency cut-off.

filterAncestry
[logical] Set to exclude samples that failed ancestry check (via check_ancestry or perSampleQC). Requires file qcdir/alg.fail-ancestry.IDs (automatically created by perSampleQC if do.check_ancestry set to TRUE).

filterRelated
[logical] Set to exclude samples that failed relatedness check (via check_relatedness or perSampleQC). Requires file qcdir/alg.fail-IBD.IDs (automatically created by perSampleQC if do.check_relatedness set to TRUE).

filterSex
[logical] Set to exclude samples that failed the sex check (via check_sex or perSampleQC). Requires file qcdir/alg.fail-sexcheck.IDs (automatically created by perSampleQC if do.check_sex set to TRUE).

filterHeterozygosity
[logical] Set to exclude samples that failed check for outlying heterozygosity rates (via check_heterozygosity_and_missingness or perSampleQC). Requires file qcdir/alg.fail-het.IDs (automatically created by perSampleQC if do.check_heterozygosity_and_missingness set to TRUE).

filterSampleMissingness
[logical] Set to exclude samples that failed check for excessive missing genotype rates (via check_heterozygosity_and_missingness or perSampleQC). Requires file qcdir/alg.fail-imiss.IDs (automatically created by perSampleQC if do.check_heterozygosity_and_missingness set to TRUE).

filterSNPMissingness
[logical] Set to exclude markers that have excessive missing rates across samples (via check_snp_missingness or perMarkerQC). Requires lmissTh to be set.

filterHWE
[logical] Set to exclude markers that fail HWE exact test (via check_hwe or perMarkerQC). Requires hweTh to be set.

filterMAF
[logical] Set to exclude markers that fail minor allele freuqency or minor allele count threshold (via check_maf or perMarkerQC). Requires mafTh or macTh to be set.

path2plink
[character] Absolute path to where external plink software https://www.cog-genomics.org/plink/1.9/ can be found. If not provided, assumed that PATH set-up works and plink will be found by system("plink").

verbose
[logical] If TRUE, progress info is printed to standard out.

showPlinkOutput
[logical] If TRUE, plink log and error messages are printed to standard out.

Value

names [list] with i) passIDs, containing a [data.frame] with family [FID] and individual [IID] IDs of samples that pass the QC, ii) failIDs, containing a [data.frame] with family [FID] and individual [IID] IDs of samples that fail the QC.