plinkQC: Genotype Quality Control with 'PLINK'

Genotyping arrays enable the direct measurement of an individuals genotype at thousands of markers. 'plinkQC' facilitates genotype quality control for genetic association studies as described by Anderson and colleagues (2010) doi:10.1038/nprot.2010.116 . It makes 'PLINK' basic statistics (e.g. missing genotyping rates per individual, allele frequencies per genetic marker) and relationship functions accessible from 'R' and generates a per-individual and per-marker quality control report. Individuals and markers that fail the quality control can subsequently be removed to generate a new, clean dataset. Removal of individuals based on relationship status is optimised to retain as many individuals as possible in the study. Additionally, there is a pre-trained classifier to predict genomic ancestry of human samples.

See also

Author

Maintainer: Hannah Meyer hannah.v.meyer@gmail.com (ORCID)

Other contributors:

Caroline Walter carowalter26@hotmail.com [contributor]

Maha Syed maha.s.syed@gmail.com [contributor]