Run PLINK sexcheck

Run plink --sexcheck to calculate the heterozygosity rate across X-chromosomal variants.

run_check_sex(indir, name, qcdir = indir, verbose = FALSE,
  path2plink = NULL, showPlinkOutput = TRUE)

Arguments

indir
[character] /path/to/directory containing the basic PLINK data files name.bim, name.bed, name.fam files.

name
[character] Prefix of PLINK files, i.e. name.bed, name.bim, name.fam.

qcdir
[character] /path/to/directory to save name.sexcheck as returned by plink --check-sex. User needs writing permission to qcdir. Per default qcdir=indir.

verbose
[logical] If TRUE, progress info is printed to standard out.

path2plink
[character] Absolute path to directory where external plink software https://www.cog-genomics.org/plink/1.9/ can be found, i.e. plink should be accessible as path2plink/plink -h. If not provided, assumed that PATH set-up works and plink will be found by system("plink").

showPlinkOutput
[logical] If TRUE, plink log and error messages are printed to standard out.

Details

Both run_check_sex and its evaluation evaluate_check_sex can simply be invoked by check_sex.

Examples

indir <- system.file("extdata", package="plinkQC")
name <- 'data'
qcdir <- tempdir()
# the following code is not run on package build, as the path2plink on the
# user system is not known.
# NOT RUN {
run <- run_check_sex(indir=indir, qcdir=qcdir, name=name)
# }