Quality control for all individuals in plink-dataset

perIndividualQC checks the samples in the plink dataset for their total missingness and heterozygosity rates, the concordance of their assigned sex to their SNP sex, their relatedness to other study individuals and their genetic ancestry.

perIndividualQC(qcdir, alg, do.run_check_sex = TRUE,
  do.evaluate_check_sex = TRUE, maleTh = 0.8, femaleTh = 0.2,
  fixMixup = FALSE, externalSex = NULL, externalMale = "M",
  externalSexSex = "Sex", externalSexID = "IID",
  externalFemale = "F", do.run_check_het_and_miss = TRUE,
  do.evaluate_check_het_and_miss = TRUE, imissTh = 0.03, hetTh = 3,
  do.run_check_relatedness = TRUE,
  do.evaluate_check_relatedness = TRUE, highIBDTh = 0.1875,
  do.check_ancestry = TRUE, prefixMergedDataset, europeanTh = 1.5,
  refSamples = NULL, refColors = NULL, refSamplesFile = NULL,
  refColorsFile = NULL, refSamplesIID = "IID", refSamplesPop = "Pop",
  refColorsColor = "Color", refColorsPop = "Pop",
  studyColor = "#2c7bb6", interactive = FALSE, verbose = TRUE,
  path2plink = NULL, showPlinkOutput = TRUE)

Arguments

qcdir
[character] /path/to/directory/with/QC/results containing the basic data files alg.bim, alg.bed, alg.fam files. In addition, if do.evaluate_[analysis] is set to TRUE and do.run_[analysis] is FALSE, it expects the analysis-specific plink output files in qcdir i.e. do.check_sex expects alg.sexcheck, do.evaluate_check_het_and_miss expects alg.het and alg.imiss, do.evauluate_check_relatedness expects alg.genome and alg.imiss and do.check_ancestry expects prefixMergeData.eigenvec. If these files are not present perIndividualQC will fail with missing file error. Setting do.run_[analysis] TRUE will execute the checks and create the required files. User needs writing permission to qcdir.

alg
[character] Prefix of plink files, i.e. alg.bed, alg.bim, alg.fam.

do.run_check_sex
[logical] If TRUE, run run_check_sex

do.evaluate_check_sex
[logical] If TRUE, run evaluate_check_sex

maleTh
[double] Threshold of X-chromosomal heterozygosity rate for males.

femaleTh
[double] Threshold of X-chromosomal heterozygosity rate for females.

fixMixup
[logical] Should PEDSEX of individuals with mismatch between PEDSEX and Sex, with Sex==SNPSEX automatically corrected: this will directly change the alg.bim/.bed/.fam files!

externalSex
[data.frame, optional] Dataframe with sample IDs [externalSexID] and sex [externalSexSex] to double check if external and PEDSEX data (often processed at different centers) match.

externalMale
[integer/character] Identifier for 'male' in externalSex.

externalSexSex
[character] Column identifier for column containing sex information in externalSex.

externalSexID
[character] Column identifier for column containing ID information in externalSex.

externalFemale
[integer/character] Identifier for 'female' in externalSex.

do.run_check_het_and_miss
[logical] If TRUE, run run_check_heterozygosity and run_check_missingness

do.evaluate_check_het_and_miss
[logical] If TRUE, run evaluate_check_het_and_miss.

imissTh
[double] Threshold for acceptable missing genotype rate in any individual

hetTh
[double] Threshold for acceptable deviation from mean heterozygosity in any individual. Expressed as multiples of standard deviation of heterozygosity (het), i.e. individuals outside mean(het) +/- hetTh*sd(het) will be returned as failing heterozygosity check.

do.run_check_relatedness
[logical] If TRUE, run run_check_relatedness.

do.evaluate_check_relatedness
[logical] If TRUE, run evaluate_check_relatedness.

highIBDTh
[double] Threshold for acceptable proportion of IBD between pair of individuals.

do.check_ancestry
[logical] If TRUE, run check_ancestry.

prefixMergedDataset
[character] Prefix of merged dataset (study and reference samples) used in plink --pca, resulting in prefixMergedDataset.eigenvec

europeanTh
[double] Scaling factor of radius to be drawn around center of reference European samples, with study samples inside this radius considered to be of European descent and samples outside this radius of non-European descent. The radius is computed as the maximum Euclidean distance of reference Europeans to the centre of reference European samples.

refSamples
[data.frame] Dataframe with sample identifiers [refSamplesIID] corresponding to IIDs in prefixMergedDataset.eigenvec and population identifier [refSamplesPop] corresponding to population IDs [refColorsPop] in refColorsfile/refColors. Either refSamples or refSamplesFile have to be specified.

refColors
[data.frame, optional] Dataframe with population IDs in column [refColorsPop] and corresponding color-code for PCA plot in column [refColorsColor]. If not provided and is.null(refColorsFile) default colors are used.

refSamplesFile
[character] /path/to/File/with/reference samples. Needs columns with sample identifiers [refSamplesIID] corresponding to IIDs in prefixMergedDataset.eigenvec and population identifier [refSamplesPop] corresponding to population IDs [refColorsPop] in refColorsfile/refColors.

refColorsFile
[character, optional] /path/to/File/with/Population/Colors cotaining population IDs in column [refColorsPop] and corresponding color-code for PCA plot in column [refColorsColor].If not provided and is.null(refColors) default colors for are used.

refSamplesIID
[character] Column name of reference sample IDs in refSamples/refSamplesFile.

refSamplesPop
[character] Column name of reference sample population IDs in refSamples/refSamplesFile.

refColorsColor
[character] Column name of population colors in refColors/refColorsFile

refColorsPop
[character] Column name of reference sample population IDs in refColors/refColorsFile.

studyColor
[character] Color to be used for study population if plot is TRUE.

interactive
[logical] Should plots be shown interactively? When choosing this option, make sure you have X-forwarding/graphical interface available for interactive plotting. Alternatively, set interactive=FALSE and save the returned plot object (p_sampleQC) via ggplot2::ggsave(p=p_sampleQC, other_arguments) or pdf(outfile) print(p_sampleQC) dev.off(). If TRUE, i) depicts the X-chromosomal heterozygosity (SNPSEX) of the samples split by their PEDSEX (if do.evaluate_check_sex is TRUE), ii) creates a scatter plot with samples' missingness rates on x-axis and their heterozygosity rates on the y-axis (if do.evaluate_check_het_and_miss is TRUE), iii) depicts all pair-wise IBD-estimates as histogram (if do.evaluate_check_relatedness is TRUE) and iv) creates a scatter plot of PC1 versus PC2 color-coded for samples of reference populations and study population (if do.check_ancestry is set to TRUE).

verbose
[logical] If TRUE, progress info is printed to standard out.

path2plink
[character] Absolute path to where external plink software https://www.cog-genomics.org/plink/1.9/ can be found. If not provided, assumed that PATH set-up works and plink will be found by system("plink").

showPlinkOutput
[logical] If TRUE, plink log and error messages are printed to standard out.

Value

Named [list] with i) fail_list, a named [list] with 1. sample_missingness containing a [vector] with sample IIDs failing the missingness threshold imissTh, 2. highIBD containing a [vector] with sample IIDs failing the relatedness threshold highIBDTh, 3. outlying_heterozygosity containing a [vector] with sample IIDs failing the heterozygosity threshold hetTh, 4. mismatched_sex containing a [vector] with the sample IIDs failing the sexcheck based on SNPSEX and femaleTh/maleTh and 5. ancestry containing a vector with sample IIDs failing the ancestry check based on europeanTh and ii) p_sampleQC, a ggplot2-object 'containing' a sub-panelled plot with the QC-plots of check_sex, check_het_and_miss, check_relatedness and check_ancestry, which can be shown by print(p_sampleQC). List entries contain NULL if that specific check was not chosen.

Details

perIndividualQC wraps around the individual QC functions check_sex, check_het_and_miss, check_relatedness and check_ancestry. For details on the parameters and outputs, check these function documentations. For detailed output for fail IIDs (instead of simple IID lists), run each function individually.

Examples

package.dir <- find.package('plinkQC')
qcdir <- file.path(package.dir, 'extdata')
alg <- "data"
# All quality control checks
fail_individuals <- perIndividualQC(qcdir=qcdir, alg=alg,
refSamplesFile=paste(qcdir, "/HapMap_ID2Pop.txt",sep=""),
refColorsFile=paste(qcdir, "/HapMap_PopColors.txt", sep=""),
prefixMergedDataset="data.HapMapIII", interactive=FALSE, verbose=FALSE,
do.run_check_het_and_miss=FALSE, do.run_check_relatedness=FALSE,
do.run_check_sex=FALSE)

#> Error in evaluate_check_sex(qcdir = qcdir, alg = alg, maleTh = maleTh,     femaleTh = femaleTh, externalSex = externalSex, externalMale = externalMale,     externalFemale = externalFemale, externalSexSex = externalSexSex,     externalSexID = externalSexID, verbose = verbose, path2plink = path2plink,     showPlinkOutput = showPlinkOutput, fixMixup = fixMixup, interactive = FALSE): plink --check-sex results file: /Users/hannah/RPackagesDevel/plinkQC/extdata/data.sexcheck does not exist.

# Only check sex and missingness/heterozygosity
fail_sex_het_miss <- perIndividualQC(qcdir=qcdir, alg=alg,
do.check_ancestry=FALSE, do.evaluate_check_relatedness=FALSE,
interactive=FALSE, verbose=FALSE, do.run_check_het_and_miss=FALSE,
do.run_check_relatedness=FALSE, do.run_check_sex=FALSE)

#> Error in evaluate_check_sex(qcdir = qcdir, alg = alg, maleTh = maleTh,     femaleTh = femaleTh, externalSex = externalSex, externalMale = externalMale,     externalFemale = externalFemale, externalSexSex = externalSexSex,     externalSexID = externalSexID, verbose = verbose, path2plink = path2plink,     showPlinkOutput = showPlinkOutput, fixMixup = fixMixup, interactive = FALSE): plink --check-sex results file: /Users/hannah/RPackagesDevel/plinkQC/extdata/data.sexcheck does not exist.